+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.2.3 (2025-11-30)
+
+- [x] (feature) Normalized string output using the `strx` package.
+
+- [x] (fix) Fixed warning on `Expr.map_elements` in `parse_commerical_invoice`
+
+- [x] (docs) Updated README document
+
 ## v0.1.3 (2025-10-02)
 
 - [x] (docs) Updated README